Solution to Find the Median of an Array

+++ Find the Median of an Array@@676c02002dbc8c3e7a280b8c/code.c
@@ -26,7 +26,7 @@ int main() {
         printf("%d\n", arr[n / 2]);
     } else {
         // Even size: average of two middle elements
-        float median = (arr[n / 2 - 1] + arr[n / 2]) / 2.0;
+        int median = (arr[n / 2 - 1] + arr[n / 2]) / 2.0;
         printf("%d\n", median);
     }
